feat(app): promote new navigational structure to default view
Invert the meaning of the hierarchyReorganization feature flag. Unless the flag is set to true, the
new app ui layout is the default.
