+++ app-testing/tests/app_settings_v5dot1_test.py
@@ -32,7 +32,7 @@ def test_app_settings_v5dot1(
     # app should look on localhost for robots
     os.environ["OT_APP_DISCOVERY__CANDIDATES"] = "localhost"
     # app should use the __DEV__ Hierarchy Reorganization
-    os.environ["OT_APP_DEV_INTERNAL__hierarchyReorganization"] = "true"
+    os.environ["OT_APP_DEV_INTERNAL__hierarchyReorganization"] = "false"
     # Start chromedriver with our options and use the
     # context manager to ensure it quits.
     with webdriver.Chrome(options=chrome_options) as driver:
