+++ app/src/assets/localization/en/app_settings.json
@@ -74,5 +74,8 @@
   "clear_robots_description": "Clear the list of unavailable robots on the Devices page. This action cannot be undone.",
   "clear_robots_button": "Clear unavailable robots list",
   "enable_dev_tools": "Enable Developer Tools",
-  "enable_dev_tools_description": "Enabling this setting opens Developer Tools on app launch, enables additional logging and gives access to feature flags."
+  "enable_dev_tools_description": "Enabling this setting opens Developer Tools on app launch, enables additional logging and gives access to feature flags.",
+  "__dev_internal__hierarchyReorganization": "revert to legacy hierarchy organization",
+  "__dev_internal__allPipetteConfig": "All Pipette Config",
+  "__dev_internal__enableBundleUpload": "Enable Bundle Upload"
 }
