+++ app/src/organisms/AppSettings/FeatureFlags.tsx
@@ -16,13 +16,14 @@ import * as Config from '../../redux/config'
 import type { DevInternalFlag } from '../../redux/config/types'
 import { Divider } from '../../atoms/structure'
 import { ToggleButton } from '../../atoms/Buttons'
+import { useTranslation } from 'react-i18next'
 
 export function FeatureFlags(): JSX.Element {
+  const { t } = useTranslation('app_settings')
   const devInternalFlags = useSelector(Config.getFeatureFlags)
   const dispatch = useDispatch<Dispatch>()
   const toggleDevInternalFlag = (flag: DevInternalFlag): unknown =>
     dispatch(Config.toggleDevInternalFlag(flag))
-  console.log(Config.DEV_INTERNAL_FLAGS)
 
   return (
     <Box
@@ -38,7 +39,7 @@ export function FeatureFlags(): JSX.Element {
             key={flag}
           >
             <Text css={TYPOGRAPHY.h3SemiBold} id={`FeatureFlags_${flag}_text`}>
-              {startCase(flag)}
+              {t(`__dev_internal__${flag}`)}
             </Text>
             <ToggleButton
               label={`${flag}-toggle`}
