+++ app/src/organisms/Devices/HeaterShakerWizard/index.tsx
@@ -21,7 +21,7 @@ import {
   useHoverTooltip,
 } from '@opentrons/components'
 
-import type { NextGenRouteParams } from '../../../App/types'
+import type { NavRouteParams } from '../../../App/types'
 import type { HeaterShakerModule } from '../../../redux/modules/types'
 import type { ProtocolModuleInfo } from '../../ProtocolSetup/utils/getProtocolModulesInfo'
 import type { ThermalAdapterName } from '@opentrons/shared-data'
@@ -37,7 +37,7 @@ export const HeaterShakerWizard = (
   const { onCloseClick, moduleFromProtocol } = props
   const { t } = useTranslation(['heater_shaker', 'shared'])
   const [currentPage, setCurrentPage] = React.useState(0)
-  const { robotName } = useParams<NextGenRouteParams>()
+  const { robotName } = useParams<NavRouteParams>()
   const attachedModules = useAttachedModules(robotName)
   const [targetProps, tooltipProps] = useHoverTooltip()
   const heaterShaker =
