+++ app/src/pages/Devices/DeviceDetails/index.tsx
@@ -19,10 +19,10 @@ import { PipettesAndModules } from '../../../organisms/Devices/PipettesAndModule
 import { RecentProtocolRuns } from '../../../organisms/Devices/RecentProtocolRuns'
 import { RobotOverview } from '../../../organisms/Devices/RobotOverview'
 
-import type { NextGenRouteParams } from '../../../App/types'
+import type { NavRouteParams } from '../../../App/types'
 
 export function DeviceDetails(): JSX.Element | null {
-  const { robotName } = useParams<NextGenRouteParams>()
+  const { robotName } = useParams<NavRouteParams>()
 
   const robot = useRobot(robotName)
